1) sign-up: - reworked the error handling - added more comments
2) edit/save users
- added error handling + user feedback

+++ Typemaster/php/deleteUser.php
@@ -4,10 +4,15 @@
 require_once("db_connection.php");
 $conn = connection();
 
+if(!$conn) {
+    $errMsg = "Connection to the Webserver failed";
+    print $errMsg;
+    exit();
+}
+
 // get Request from JS Promise - decode JS Object
 // the JS Object only contains the id
 $id = json_decode(file_get_contents("php://input"));
-
 
 // sql command - delete entry in the users table -> the row with the specified id
 $sql = "DELETE FROM users WHERE id = "."'".$id."'";
